refactor: [debugger] Code optimization
Log:
