+++ src/plugins/debugger/interface/appoutputpane.h
@@ -6,33 +6,17 @@
 #define APPOUTPUTPANE_H
 
 #include "debuggerglobals.h"
+#include "common/widget/outputpane.h"
 
 #include <QPlainTextEdit>
 
 class OutputWindowPrivate;
-class AppOutputPane : public QPlainTextEdit
+class AppOutputPane : public OutputPane
 {
     Q_OBJECT
 public:
     AppOutputPane();
     ~AppOutputPane() override;
-
-    void clearContents();
-
-    bool isScrollbarAtBottom() const;
-    QString doNewlineEnforcement(const QString &out);
-    void scrollToBottom();
-
-protected:
-    void contextMenuEvent(QContextMenuEvent * event) override;
-
-public slots:
-    void appendText(const QString &text, OutputPane::OutputFormat format);
-
-private:
-    void appendText(const QString &text, const QTextCharFormat &format = QTextCharFormat());
-    QList<QAction*> actionFactory();
-    OutputWindowPrivate *d = nullptr;
 };
 
 #endif   // APPOUTPUTPANE_H
